[RF] tolerance is floating, not integer
Fixes https://github.com/root-project/root/issues/20894

Bug introduced in https://github.com/root-project/root/commit/a3c91214b1881ed6070533a970762b7a35863934

+++ roofit/roostats/src/ProfileLikelihoodCalculator.cxx
@@ -160,7 +160,7 @@ RooFit::OwningPtr<RooFitResult> ProfileLikelihoodCalculator::DoMinimizeNLL(RooAb
    const char * minimAlgo = ROOT::Math::MinimizerOptions::DefaultMinimizerAlgo().c_str();
    int strategy = ROOT::Math::MinimizerOptions::DefaultStrategy();
    int level = ROOT::Math::MinimizerOptions::DefaultPrintLevel() -1;// RooFit level starts from  -1
-   int tolerance = ROOT::Math::MinimizerOptions::DefaultTolerance();
+   double tolerance = ROOT::Math::MinimizerOptions::DefaultTolerance();
    // do global fit and store fit result for further use
 
    const auto& config = GetGlobalRooStatsConfig();
